+++ pkg/dist-src/components/controls/index.stories.js
@@ -1,4 +1,4 @@
-import { define } from 'hybrids';
+import { html, define } from 'hybrids';
 import Controls from ".//index.js";
 import Fieldset from "../fieldset/index.js";
 import Field from "../field/index.js";
@@ -8,19 +8,39 @@ define('qs-field', Field);
 export default {
   title: 'Controls|<qs-controls>'
 };
-export const Demo = () => `
-<qs-controls>
-  <qs-fieldset legend="Checkbox Sample">
-    <label><input type="checkbox" /> first</label>
-    <label><input type="checkbox" /> second</label>
-  </qs-fieldset>
-  <qs-fieldset legend="Radio Sample">
-    <label><input type="radio" /> first</label>
-    <label><input type="radio" /> second</label>
-  </qs-fieldset>
-  <qs-field label="Sample Field" label-pos="top">
-    <input type="color" />
-  </qs-field>
-  <button type="button">Click me</button>
-</qs-controls>
-`;
+define('storybook-controls', {
+  render: () => html`
+    <style>
+      qs-controls {
+        --form-font: monospace;
+        --form-font-size: 1rem;
+        --group-border: 4px solid gainsboro;
+        --group-padding: 0.75em;
+        --field-label-font-size: 2em;
+      }
+    </style>
+    <qs-controls>
+      <qs-fieldset group="Checkbox Sample">
+        <qs-field label-for="first" label="first" label-pos="right"
+          ><input type="checkbox" />
+        </qs-field>
+        <qs-field label-for="second" label="second" label-pos="right"
+          ><input type="checkbox" />
+        </qs-field>
+      </qs-fieldset>
+      <qs-fieldset group="Radio Sample">
+        <qs-field label-for="first" label="first" label-pos="right"
+          ><input type="radio" />
+        </qs-field>
+        <qs-field label-for="second" label="second" label-pos="right"
+          ><input type="radio" />
+        </qs-field>
+      </qs-fieldset>
+      <qs-field label="Color" label-for="color">
+        <input type="color" value="#f00000" />
+      </qs-field>
+      <button type="button">Click me</button>
+    </qs-controls>
+  `
+});
+export const Demo = () => `<storybook-controls></storybook-controls>`;