+++ pkg/dist-src/components/data/index.stories.js
@@ -1,7 +1,12 @@
-import { define } from 'hybrids';
+import { html, define } from 'hybrids';
 import Data from ".//index.js";
 define('qs-data', Data);
 export default {
   title: 'Display|Data/<qs-data>'
 };
-export const Demo = () => document.createElement('qs-data');
+define('storybook-data', {
+  render: () => html`
+      <qs-data></qs-data>
+    `
+});
+export const Demo = () => `<storybook-data></storybook-data>`;