+++ pkg/dist-src/components/fieldset/index.stories.js
@@ -1,12 +1,15 @@
-import { define } from 'hybrids';
+import { html, define } from 'hybrids';
 import Fieldset from ".//index.js";
 define('qs-fieldset', Fieldset);
 export default {
   title: 'Controls|<qs-fieldset>'
 };
-export const Demo = () => `
-  <qs-fieldset legend="Radio Sample">
-    <label><input type="radio" /> first</label>
-    <label><input type="radio" /> second</label>
-  </qs-fieldset>
-`;
+define('storybook-fieldset', {
+  render: () => html`
+    <qs-fieldset group="Radio Sample">
+      <label><input type="radio" /> first</label>
+      <label><input type="radio" /> second</label>
+    </qs-fieldset>
+  `
+});
+export const Demo = () => `<storybook-fieldset></storybook-fieldset>`;